new-workshops/pyproject.toml
[project]
name = "neo4j-graphrag-workshop"
version = "1.0.0"
description = "Neo4j and AWS Bedrock GraphRAG Workshop"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    # Core AWS SDK
    "boto3>=1.35.0",
    "botocore>=1.35.0",

    # Strands Agents SDK (AWS open-source agent framework)
    # See: https://strandsagents.com
    "strands-agents>=0.2.0",
    "strands-agents-tools>=0.2.0",

    # LangChain integration for Neo4j tools
    "langchain>=0.3.0",
    "langchain-aws>=0.2.0",
    "langchain-community>=0.3.0",

    # Neo4j
    "neo4j>=5.0.0",
    # Using local fork with Bedrock support
    # After testing, replace with: "neo4j-graphrag[bedrock]>=1.10.0"
    "neo4j-graphrag[bedrock] @ file:///Users/ryanknight/projects/neo4j-graphrag-python",

    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # Jupyter notebooks
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["solutions"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
